Relocate agent (manual changes)
