+++ .golangci.yml
@@ -123,3 +123,6 @@ issues:
   # Don't hide multiple issues that belong to one class since GitHub annotations can handle them all nicely.
   max-issues-per-linter: 0
   max-same-issues: 0
+
+  exclude-dirs:
+    - internal/agent
